@@ -1,7 +1,7 @@
 [package]
 name = "decancer_java"
 version = "2.0.2"
-edition = "2021"
+edition = "2024"
 publish = false
 
 [lib]
