@@ -1,7 +1,7 @@
 [package]
 name = "decancer"
-version = "3.3.3"
-edition = "2021"
+version = "4.0.0"
+edition = "2024"
 authors = ["null (https://github.com/null8626)"]
 description = "A library that removes common unicode confusables/homoglyphs from strings."
 readme = "README.md"
@@ -10,12 +10,11 @@ license = "MIT"
 keywords = ["security", "unicode", "homoglyphs", "moderation", "confusables"]
 categories = ["parsing", "text-processing", "value-formatting"]
 exclude = ["rustfmt.toml"]
-rust-version = "1.65.0"
+rust-version = "1.85.0"
 
 [dependencies]
 serde = { version = "1", default-features = false, optional = true }
 regex = { version = "1", default-features = false, features = ["std", "perf"], optional = true }
-lazy_static = { version = "1", optional = true }
 
 [lints.clippy]
 all = { level = "warn", priority = -1 }
@@ -38,7 +37,7 @@ default = ["options", "separators", "leetspeak"]
 serde = ["dep:serde"]
 options = []
 separators = []
-leetspeak = ["regex", "lazy_static"]
+leetspeak = ["regex"]
 
 [dev-dependencies]
 proptest = { version = "1", default-features = false, features = ["std"] }
